requirements.txt
groq
pandas
python-dotenv
streamlit
fpdf
google-play-scraper
markdown
